# swagger.yaml
...
swagger: '2.0'
info:
  title: IoT Device Repository
  termsOfService: http://swagger.io/terms/
  description: 'The IOT Device Repository provides metadata for all IOT devices managed on the platform, their device classes and the associated data structures for the user.'
  version: '1.0'
  contact:
    name: sh20xyqi
    email: sh20xyqi@studserv.uni-leipzig.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - https
host: api.sepl.infai.org
basePath: /iot-device-repo
produces:
  - text/plain
...
paths:
...
  /deviceInstance:
    post:
      tags:
        - deviceInstanceApi
      summary:
        Add a new device instance.
      description:
        ''
      produces:
        - application/json
      parameters:
      - in: body
        name: body
        description: Device instance object that needs to be added to the database.
        required: true
        schema:
          $ref: '#/definitions/DeviceInstance'
      responses:
        200:
          description: Returns the instanciated device instance.
          schema:
            $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - JSON encoding failed; DeviceInstance is not valid;  DeviceInstance is not consistent; DeviceInstance can not be instanciated.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /deviceInstance/{id}:
    post:
      tags:
        - deviceInstanceApi
      summary:
        Updates device instance by Id.
      description:
        ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the the to be updated device instance.
          required: true
          type: string
        - in: body
          name: body
          description: Device instance object that needs to be updated.
          required: true
          schema:
            $ref: '#/definitions/DeviceInstance'
      responses:
        200:
          description: Returns that the device instance that has been modified.
          schema:
            $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceInstance is not valid;  DeviceInstance is not consistent; DeviceInstance can not be instanciated.
          schema:
            $ref: '#/definitions/ErrorMessage'
    get:
      tags:
        - deviceInstanceApi
      summary:
        Returns device instance by Id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the the to be returned device instance.
          required: true
          type: string
      responses:
        200:
          description: Returns the device instance behind the passed ID.
          schema:
            $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceInstance is not valid;  DeviceInstance is not consistent; DeviceInstance can not be instanciated.
          schema:
            $ref: '#/definitions/ErrorMessage'
    delete:
      tags:
        - deviceInstanceApi
      summary:
        Deletes device instance by Id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the the to be deleted device instance.
          required: true
          type: string
      responses:
        200:
          description: Deletes the device instance behind the passed ID.
          schema:
            $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceInstance can not be found in database; DeviceInstance can not be deleted in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /deviceInstances/{limit}/{offset}:
    get:
      tags:
        - deviceInstanceApi
      summary:
        Returns device instances within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns  a sorted list of device instances.
          schema:
            type: array
            items:
              $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; a list of device instances can not be returned.
          schema:
            $ref: '#/definitions/ErrorMessage'
  #Everything that belongs to deviceInstanceApi/guiApi.go
  /ui/deviceInstance/resourceSkeleton/{deviceTypeId}:
    get:
      tags:
        - deviceInstanceApi
      summary:
        Returns an empty instance by using the device type id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: deviceTypeId
          in: path
          description: Identifies the device type by id and creates an to be returned empty instance.
          required: true
          type: string
      responses:
        200:
          description: Returns the empty instance.
          schema:
            $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceTypeId can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /url_to_devices/{device}:
    get:
      tags:
        - deviceInstanceApi
      summary:
        Returns the device service entity by using the device.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: device
          in: path
          description: Identifies the device in the database.
          required: true
          type: string
      responses:
        200:
          description: Returns the device service entity.
          schema:
            $ref: '#/definitions/DeviceServiceEntity'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceServiceEntity can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /bydevicetype/deviceInstances/{devicetype}:
    get:
      tags:
        - deviceInstanceApi
      summary:
        Returns a sorted list containing all device instances.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: devicetype
          in: path
          description: Identifies the to be returned list of device instances by using the passed device type.
          required: true
          type: string
      responses:
        200:
          description: Returns the specific sorted list.
          schema:
            type: array
            items:
              $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceType can not be found in database; DeviceInstance can not be found in database when using its Id.
          schema:
            $ref: '#/definitions/ErrorMessage'
  #Everything that belongs to deviceTypeApi
  #Everything that belongs to deviceTypeApi/crudApi.go
...
definitions:
...
  DeviceServiceEntity:
    type: object
    properties:
      device:
        type: object
        format: '#/definitions/DeviceInstance'
      services:
        type: array
        items:
          $ref: '#/definitions/ShortService'
    required:
      - device
      - services
  ShortService:
    type: object
    properties:
      id:
        type: string
        format: string
      service_type:
        type: string
        format: string
      url:
        type: string
        format: string
    required:
      - id
      - service_type
      - url
...
  DeviceInstance:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      device_type:
        type: string
        format: string
      config:
        type: array
        items:
          $ref: '#/definitions/ConfigField'
      uri:
        type: string
        format: string
    required:
      - id
      - name
      - device_type
      - config
      - uri
...
  ConfigField:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      value:
        type: string
        format: string
    required:
      - id
      - name
      - value
...
  ErrorMessage:
    type: object
    properties:
      status_code:
        type: integer
        format: int32
      message:
        type: string
        format: string
      error_code:
        type: string
        format: string
      detail:
        type: array
        items:
          type: string
    required:
      - status_code
      - message
      - error_code
      - detail
...